Refactor change responsability fetch from vuex
